Made changes to GreetingService updateGreeting to allow adding new language to the DB if it doesn't already exist.

+++ src/main/java/com/keyin/hello/GreetingController.java
@@ -34,7 +34,7 @@ public class GreetingController {
     }
 
     @PutMapping("greeting/{index}")
-    public Greeting updateGreeting(@PathVariable Integer index, @RequestBody Greeting updatedGreeting) {
+    public Greeting updateGreeting(@PathVariable long index, @RequestBody Greeting updatedGreeting) {
         return greetingService.updateGreeting(index, updatedGreeting);
     }
